App/clientstream/requirements.txt
#dotenv==0.0.5
#socketio==0.2.1
python-dotenv==0.20.0
python-socketio==5.7.2
